Passage Taille occupée en long (avant c'était int * int) donc on pouvait déborder

+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/tools/Photos_Outils.java
@@ -701,7 +701,7 @@ public class Photos_Outils {
     public long getPhotoDiskUsage(ImageLocation baseImageLocation, ImageType inImageType) {
         //if (BuildConfig.DEBUG) Log.d(LOG_TAG, "Photos_Outils() - getPhotoDiskUsage() - baseImageLocation : "+baseImageLocation.name());
         //if (BuildConfig.DEBUG) Log.d(LOG_TAG, "Photos_Outils() - getPhotoDiskUsage() - inImageType : "+inImageType.name());
-        return getImageCountInFolder(baseImageLocation, inImageType) * getTailleMoyImageUnitaire(inImageType);
+        return (long)getImageCountInFolder(baseImageLocation, inImageType) * (long)getTailleMoyImageUnitaire(inImageType);
     }
